add seal sampler
