--- examples/pytorch/ogb/seal_ogbl/models.py
@@ -1,79 +1,0 @@
-import math
-import torch
-from torch.nn import (ModuleList, Linear, Conv1d, MaxPool1d, Embedding)
-import torch.nn.functional as F
-from dgl.nn import GraphConv, SortPooling
-
-
-# An end-to-end deep learning architecture for graph classification, AAAI-18.
-class DGCNN(torch.nn.Module):
-    def __init__(self, hidden_channels, num_layers, max_z, k=0.6, train_dataset=None, 
-                 dynamic_train=False, GNN=GraphConv, use_feature=False):
-        super(DGCNN, self).__init__()
-
-        self.use_feature = use_feature
-
-        if k <= 1:  # Transform percentile to number.
-            if dynamic_train:
-                sampled_indices = range(1000)
-            else:
-                sampled_indices = range(len(train_dataset))
-            num_nodes = sorted([train_dataset[i][0].num_nodes() for i in sampled_indices])
-            k = num_nodes[int(math.ceil(k * len(num_nodes))) - 1]
-            k = max(10, k)
-        self.k = int(k)
-        self.sort_pool = SortPooling(k=self.k)
-
-        self.max_z = max_z
-        self.z_embedding = Embedding(self.max_z, hidden_channels)
-
-        self.convs = ModuleList()
-        initial_channels = hidden_channels
-        if self.use_feature:
-            initial_channels += train_dataset.node_features.size(1)
-
-        self.convs.append(GNN(initial_channels, hidden_channels))
-        for _ in range(0, num_layers-1):
-            self.convs.append(GNN(hidden_channels, hidden_channels))
-        self.convs.append(GNN(hidden_channels, 1))
-
-        conv1d_channels = [16, 32]
-        total_latent_dim = hidden_channels * num_layers + 1
-        conv1d_kws = [total_latent_dim, 5]
-        self.conv1 = Conv1d(1, conv1d_channels[0], conv1d_kws[0],
-                            conv1d_kws[0])
-        self.maxpool1d = MaxPool1d(2, 2)
-        self.conv2 = Conv1d(conv1d_channels[0], conv1d_channels[1],
-                            conv1d_kws[1], 1)
-        dense_dim = int((self.k - 2) / 2 + 1)
-        dense_dim = (dense_dim - conv1d_kws[1] + 1) * conv1d_channels[1]
-        self.lin1 = Linear(dense_dim, 128)
-        self.lin2 = Linear(128, 1)
-
-    def forward(self, g, z, x=None, edge_weight=None):
-        z_emb = self.z_embedding(z)
-        if z_emb.ndim == 3:  # in case z has multiple integer labels
-            z_emb = z_emb.sum(dim=1)
-        if self.use_feature and x is not None:
-            x = torch.cat([z_emb, x.to(torch.float)], 1)
-        else:
-            x = z_emb
-        xs = [x]
-
-        for conv in self.convs:
-            xs += [torch.tanh(conv(g, xs[-1], edge_weight=edge_weight))]
-        x = torch.cat(xs[1:], dim=-1)
-
-        # global pooling
-        x = self.sort_pool(g, x)
-        x = x.unsqueeze(1)  # [num_graphs, 1, k * hidden]
-        x = F.relu(self.conv1(x))
-        x = self.maxpool1d(x)
-        x = F.relu(self.conv2(x))
-        x = x.view(x.size(0), -1)  # [num_graphs, dense_dim]
-
-        # MLP.
-        x = F.relu(self.lin1(x))
-        x = F.dropout(x, p=0.5, training=self.training)
-        x = self.lin2(x)
-        return x
