--- examples/pytorch/ogb/seal_ogbl/scripts/run_citation2.sh
@@ -1,12 +1,0 @@
-python main.py \
-    --dataset ogbl-citation2 \
-    --use_feature \
-    --use_edge_weight \
-    --eval_steps 1 \
-    --epochs 10 \
-    --dynamic_train \
-    --dynamic_val \
-    --dynamic_test \
-    --train_percent 2 \
-    --val_percent 1 \
-    --test_percent 1
