--- examples/pytorch/ogb/seal_ogbl/scripts/run_collab.sh
@@ -1,5 +1,0 @@
-python main.py \
-    --dataset ogbl-collab \
-    --train_percent 15 \
-    --hidden_channels 256 \
-    --use_valedges_as_input
