--- examples/pytorch/ogb/seal_ogbl/scripts/run_ppa.sh
@@ -1,11 +1,0 @@
-python main.py \
-    --dataset ogbl-ppa \
-    --use_feature \
-    --use_edge_weight \
-    --eval_steps 5 \
-    --epochs 20 \
-    --dynamic_train \
-    --dynamic_val \
-    --dynamic_test \
-    --train_percent 5 \
-    --num_workers 10
